+++ domains/certificates/Query.API/API/GranularCertificateIssuer/Startup.cs
@@ -5,6 +5,5 @@ namespace API.GranularCertificateIssuer;
 public static class Startup
 {
     public static void AddGranularCertificateIssuer(this IServiceCollection services)
-    {
-    }
+        => services.AddSingleton<IProductionCertificateRepository, ProductionCertificateRepository>();
 }
